security(server): clear smtp auth payload on cancellation
